Use emoji as logo
SecularSteve's logo is smart, but the more I think about it, the more I believe that Tent doesn't really need a logo. Removing it means reducing the neccessary HTTP requests by one, and it also results in a more uniform look, since we already use an emoji for the settings link.